# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Suv yetkazib berish API",
        "description": "Mobil ilovalar uchun suv yetkazib berish tizimining rasmiy API hujjatlari. Barcha endpointlar Uzbek tilida izohlangan bo‘lib, mijoz (user) va xodim (director/courier) ilovalari uchun mo‘ljallangan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Asosiy API server (Laravel backend ichidagi /api prefiksi bilan)"
        }
    ],
    "paths": {
        "/v1/employees/login": {
            "post": {
                "tags": [
                    "Xodim – Autentifikatsiya"
                ],
                "summary": "Direktor yoki kuryer uchun login",
                "description": "Xodim (director/courier) telefon va parol orqali tizimga kiradi. Faqat kompaniyaga biriktirilgan va aktiv bo‘lgan xodimlarga ruxsat beriladi.",
                "operationId": "61468c6dcbe5e6291f3184298d1e1199",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Xodim telefon raqami",
                                        "type": "string",
                                        "example": "+998901234567"
                                    },
                                    "password": {
                                        "description": "Xodim paroli",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kirish muvaffaqiyatli, token va user maʼlumotlari qaytarildi"
                    },
                    "401": {
                        "description": "Telefon yoki parol noto‘g‘ri"
                    },
                    "403": {
                        "description": "Hisob bloklangan, o‘chirilgan yoki kompaniya faol emas"
                    }
                }
            }
        },
        "/v1/employees/profile": {
            "get": {
                "tags": [
                    "Xodim – Profil"
                ],
                "summary": "Joriy xodim profilini olish",
                "operationId": "37f2a330b7ae740642b83137d6c8ef5d",
                "responses": {
                    "200": {
                        "description": "Xodim va uning kompaniyasi haqidagi qisqa maʼlumotlar"
                    },
                    "401": {
                        "description": "Token noto‘g‘ri yoki yuborilmagan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/employees/logout": {
            "post": {
                "tags": [
                    "Xodim – Autentifikatsiya"
                ],
                "summary": "Xodimni tizimdan chiqarish",
                "operationId": "f2ca447c93e3dca4b3d0141b1fac150e",
                "responses": {
                    "200": {
                        "description": "Sessiya yopildi, token bekor qilindi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/employees/session-check": {
            "post": {
                "tags": [
                    "Xodim – Autentifikatsiya"
                ],
                "summary": "Sessiyani tekshirish (1 qurilma siyosati)",
                "description": "Berilgan FCM token bo‘yicha boshqa foydalanuvchi sessiyalarini yopib, joriy foydalanuvchi uchun bitta aktiv qurilma qoldiradi.",
                "operationId": "6bcce05ca4ff64f3b880f53f4ef3c153",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fcm_token"
                                ],
                                "properties": {
                                    "fcm_token": {
                                        "description": "Qurilmaning Firebase Cloud Messaging tokeni",
                                        "type": "string",
                                        "example": "fcm-token-here"
                                    },
                                    "device_type": {
                                        "description": "Qurilma turi: android yoki ios",
                                        "type": "string",
                                        "example": "android",
                                        "nullable": true
                                    },
                                    "device_name": {
                                        "description": "Qurilmaning nomi",
                                        "type": "string",
                                        "example": "Samsung A54",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sessiya muvaffaqiyatli yangilandi va bog‘landi"
                    },
                    "403": {
                        "description": "Hisob bloklangan, o‘chirilgan yoki kompaniya faol emas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer token bilan autentifikatsiya. Format: \\\"Bearer {token}\\\"",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}